Extract plugin to separate file
